... and back to a single read/write process model in core
This diversion was due to infrequent and seemingly random errors
that we were seeing previously. The root cause of those errors
seems to have been an issue with using the `jdbc` `with-db-*`
methods/macros, a connection pool, and `core.async`. Some part of
that mixture wasn't as comofrtable in a multi-thread scenario as
the rest of it.

Ditching the `jdbc` `with-db-*` invocations for explicit connection
and transaction management seems to have resolved the issue.